pyproject.toml
[tool.poetry]
name = "korablibot"
version = "0.1.0"
description = "Advanced toys for real boys"
authors = ["Arist0craft <dimnktn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
python-telegram-bot = "^13.10"
django-environ = "^0.8.1"
oci = "^2.53.1"
mcipc = "^2.3.5"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pytest-django = "^4.5.2"
pytest = "^6.2.5"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="app.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
